+++ src/http/server.ts
@@ -26,7 +26,7 @@ app.register(fastifySwagger, {
   openapi: {
     openapi: '3.0.0',
     info: {
-      title: 'Event booster',
+      title: 'Plotwist',
       version: '0.1.0',
     },
     servers: [