+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- create and get lists endpoint [#18](https://github.com/plotwist-app/plotwist-backend/pull/18)
 - health check endpoint [(#16)](https://github.com/plotwist-app/plotwist-backend/pull/16)
 
 ### Changed